Resolve HEAD from the top of the ref store backend stack
Resolving "HEAD" using the files backend will only return .git/HEAD,
but this call into ref resolve is to see if HEAD needs to be updated,
but for a namespace this will never be triggerable
since .git/HEAD most likely points outside the namespace.

Strictly speaking we should be updating every HEAD in every namespace
so a ref update not via the namespaced backend updates namespace HEAD reflogs
but as the comment above says, this fixes things for the majority of cases.

The refs parent tree is walked to find the top entry,
since the bookkeeping for every refs backend to store the top one
would require implementing another method for updating the top
on all sub backends the backend uses.