Iterate through refs in both gitdir and commondir
This changes how ref caches are populated
to pass in the git directory to base iteration from as state
instead of deciding which directory top read refs from
based on whether the ref is a per-worktree ref or not.

This is necessary to avoid hacks in iteration to forceably populate refs/bisect
if it is not in the common directory,
and when namespacing is introduced this hack breaks.